Implement Uploading JSON With Test Cases
This feature provides a new panel for uploading a JSON file containing a
series of tests to run on the automaton. This panel is togglable and
also provides two buttons to test the strings: one to run all of the
tests in sequence and another to run a specific test of the user's
choosing. The user will be informed if a test passes or fails.